server/livepatch.py: run with dropped privileges
Do not use sudo as there is no need for elevated privileges.

Because Popen() doesn't run in shell by default, providing the
environment variables as part of the arguments do not work. Extend the
command() to take a dict of extra environment variables, in which case
we copy the parent env and modify it accordingly before calling
Popen().

Also, use a private PORTAGE_TMPDIR, so we do not collide with the
system's own tmpdir.